Sez-13-90-Ordinare grid dei corsi nella pagina web

+++ MyCourse/Models/Options/CoursesOptions.cs
@@ -7,7 +7,7 @@ namespace MyCourse.Models.Options
 {
     public class CoursesOptions
     {
-        public long PerPage { get; set; }
+        public int PerPage { get; set; }
         public CoursesOrderOptions Order { get; set; }
     }
